+++ server.go
@@ -13,14 +13,16 @@ import (
 	"time"
 
 	"github.com/go-kit/kit/endpoint"
+	"github.com/go-kit/kit/log/level"
 	"github.com/go-kit/kit/sd"
 	"github.com/go-kit/kit/sd/lb"
+	"github.com/go-kit/log"
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/snappy"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/prometheus/common/log"
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/common/promlog"
 	"github.com/prometheus/prometheus/prompb"
 	yaml "gopkg.in/yaml.v2"
 )
@@ -76,7 +78,11 @@ var (
 	})
 )
 
+// Module-wide `logger` variable, initialized by `setupLogging()`.
+var logger log.Logger
+
 func init() {
+	setupLogging()
 	prometheus.MustRegister(writeDuration)
 	prometheus.MustRegister(writeErrors)
 	prometheus.MustRegister(writeSamples)
@@ -87,10 +93,19 @@ func init() {
 	prometheus.MustRegister(readSamples)
 	prometheus.MustRegister(readCrateDuration)
 	prometheus.MustRegister(readCrateErrors)
+	level.Info(logger).Log("msg", "Initialized CrateDB Prometheus Adapter", "version", version)
 }
 
 // Escaping for strings for Crate.io SQL.
 var escaper = strings.NewReplacer("\\", "\\\\", "\"", "\\\"", "'", "\\'")
+
+// Set up promlog logger.
+func setupLogging() {
+	logLevel := promlog.AllowedLevel{}
+	logLevel.Set("debug")
+	logConfig := &promlog.Config{Level: &logLevel}
+	logger = promlog.New(logConfig)
+}
 
 // Escape a labelname for use in SQL as a column name.
 func escapeLabelName(s string) string {
@@ -172,11 +187,11 @@ func responseToTimeseries(data *crateReadResponse) []*prompb.TimeSeries {
 			}
 			sort.Strings(labelnames) // Sort for unittests.
 			for _, k := range labelnames {
-				ts.Labels = append(ts.Labels, &prompb.Label{Name: string(k), Value: string(metric[model.LabelName(k)])})
+				ts.Labels = append(ts.Labels, prompb.Label{Name: string(k), Value: string(metric[model.LabelName(k)])})
 			}
 			timeseries[metric.String()] = ts
 		}
-		ts.Samples = append(ts.Samples, &prompb.Sample{Value: v, Timestamp: t})
+		ts.Samples = append(ts.Samples, prompb.Sample{Value: v, Timestamp: t})
 	}
 
 	names := make([]string, 0, len(timeseries))
@@ -220,34 +235,34 @@ func (ca *crateDbPrometheusAdapter) handleRead(w http.ResponseWriter, r *http.Re
 
 	compressed, err := ioutil.ReadAll(r.Body)
 	if err != nil {
-		log.With("err", err).Error("Failed to read body.")
+		level.Error(logger).Log("msg", "Failed to read body", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	reqBuf, err := snappy.Decode(nil, compressed)
 	if err != nil {
-		log.With("err", err).Error("Failed to decompress body.")
+		level.Error(logger).Log("msg", "Failed to decompress body", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	var req prompb.ReadRequest
 	if err := proto.Unmarshal(reqBuf, &req); err != nil {
-		log.With("err", err).Error("Failed to unmarshal body.")
+		level.Error(logger).Log("msg", "Failed to unmarshal body", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	if len(req.Queries) != 1 {
-		log.Error("More than one query sent.")
+		level.Error(logger).Log("msg", "More than one query sent")
 		http.Error(w, "Can only handle one query.", http.StatusBadRequest)
 		return
 	}
 
 	result, err := ca.runQuery(req.Queries[0])
 	if err != nil {
-		log.With("err", err).Error("Failed to run select against CrateDB.")
+		level.Error(logger).Log("msg", "Failed to run select against CrateDB", "err", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -258,14 +273,14 @@ func (ca *crateDbPrometheusAdapter) handleRead(w http.ResponseWriter, r *http.Re
 	}
 	data, err := proto.Marshal(&resp)
 	if err != nil {
-		log.With("err", err).Error("Failed to marshal response.")
+		level.Error(logger).Log("msg", "Failed to marshal response", "err", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/x-protobuf")
 	if _, err := w.Write(snappy.Encode(nil, data)); err != nil {
-		log.With("err", err).Error("Failed to compress response.")
+		level.Error(logger).Log("msg", "Failed to compress response", "err", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -305,21 +320,21 @@ func (ca *crateDbPrometheusAdapter) handleWrite(w http.ResponseWriter, r *http.R
 
 	compressed, err := ioutil.ReadAll(r.Body)
 	if err != nil {
-		log.With("err", err).Error("Failed to read body")
+		level.Error(logger).Log("msg", "Failed to read body", "err", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
 	reqBuf, err := snappy.Decode(nil, compressed)
 	if err != nil {
-		log.With("err", err).Error("Failed to decompress body")
+		level.Error(logger).Log("msg", "Failed to decompress body", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	var req prompb.WriteRequest
 	if err := proto.Unmarshal(reqBuf, &req); err != nil {
-		log.With("err", err).Error("Failed to unmarshal body")
+		level.Error(logger).Log("msg", "Failed to unmarshal body", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -331,7 +346,7 @@ func (ca *crateDbPrometheusAdapter) handleWrite(w http.ResponseWriter, r *http.R
 	writeTimer.ObserveDuration()
 	if err != nil {
 		writeCrateErrors.Inc()
-		log.With("err", err).Error("Failed to write data to CrateDB.")
+		level.Error(logger).Log("msg", "Failed to write data to CrateDB", "err", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -395,7 +410,7 @@ func (c *config) toString() string {
 func (c *config) toYaml() string {
 	data, err := yaml.Marshal(c)
 	if err != nil {
-		log.Fatalf("Serialization to YAML failed: %v", err)
+		level.Error(logger).Log("msg", "Serialization to YAML failed", "err", err)
 	}
 	return string(data)
 }
@@ -403,7 +418,7 @@ func (c *config) toYaml() string {
 func loadConfig(filename string) (*config, error) {
 	conf := &config{}
 	if filename != "" {
-		log.Infof("Reading configuration from file \"%s\"", filename)
+		level.Info(logger).Log("msg", "Reading configuration from file", "filename", filename)
 		content, err := ioutil.ReadFile(filename)
 		if err != nil {
 			return nil, fmt.Errorf("reading configuration file %q failed: %v", filename, err)
@@ -413,7 +428,7 @@ func loadConfig(filename string) (*config, error) {
 			}
 		}
 	} else {
-		log.Warnf("No configuration file used, falling back to built-in configuration")
+		level.Warn(logger).Log("msg", "No configuration file used, falling back to built-in configuration")
 		item := endpointConfig{}
 		conf.Endpoints = []endpointConfig{item}
 	}
@@ -451,7 +466,7 @@ func builtinConfig() *config {
 
 func main() {
 
-	log.Infof("Starting CrateDB Prometheus Adapter %v", version)
+	level.Info(logger).Log("msg", "Starting CrateDB Prometheus Adapter", "version", version)
 
 	flag.Parse()
 
@@ -467,7 +482,7 @@ func main() {
 
 	conf, err := loadConfig(*configFile)
 	if err != nil {
-		log.Fatalf("Error loading configuration %q: %v", *configFile, err)
+		level.Error(logger).Log("msg", "Error loading configuration", "config", *configFile, "err", err)
 	}
 
 	subscriber := sd.FixedEndpointer{}
@@ -493,7 +508,8 @@ func main() {
 	http.HandleFunc("/write", ca.handleWrite)
 	http.HandleFunc("/read", ca.handleRead)
 	http.Handle("/metrics", promhttp.Handler())
-	log.With("address", *listenAddress).Info("Listening ...")
-	log.With("endpoints", conf.toString()).Info("Connecting ...")
-	log.Fatal(http.ListenAndServe(*listenAddress, nil))
+	level.Info(logger).Log("msg", "Listening ...", "address", *listenAddress)
+	level.Info(logger).Log("msg", "Connecting ...", "endpoints", conf.toString())
+	listen_error := http.ListenAndServe(*listenAddress, nil)
+	level.Info(logger).Log("msg", "Final outcome", "err", listen_error)
 }