+++ CHANGES.rst
@@ -13,6 +13,7 @@ Unreleased
   The corresponding settings to configure the maximum pool sizes
   are ``read_pool_size_max`` and ``write_pool_size_max``.
 - Use a DSN-style connection string for talking to pgx5
+- Update Prometheus libraries (client: 1.16, server: 2.46)
 
 BREAKING CHANGES
 ----------------